requirements.txt
pocketflow>=0.0.1
google-genai>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
fastapi>=0.110.0
uvicorn>=0.29.0
streamlit>=1.33.0